General Refactor.

+++ GA/src/edu/itba/ia/tp1/engine/population/selection/EliteUniversalImpl.java
@@ -21,7 +21,7 @@ public class EliteUniversalImpl implements I_SelectionAlgorithm {
 	 * @see edu.itba.ia.tp1.engine.population.manager.IPopulationAlgorithm#execute(edu.itba.ia.tp1.engine.population.Population)
 	 */
 	@Override
-	public Population execute(Population population, I_Aptitude aptitude,Integer nIndividuals) {
+	public Population execute(Population population, I_Aptitude aptitude, Long nIndividuals) {
 		return null;
 	}
